Fix sigmoid bug

+++ multimodal/src/autogluon/multimodal/data/preprocess_dataframe.py
@@ -441,7 +441,7 @@ class MultiModalFeaturePreprocessor(TransformerMixin, BaseEstimator):
         Predicted labels ready to compute metric scores.
         """
         if loss_func is not None and isinstance(loss_func, torch.nn.BCEWithLogitsLoss):
-            y_pred = torch.sigmoid(y_pred)
+            y_pred = torch.sigmoid(y_pred.float())
 
         y_pred = y_pred.detach().cpu().float().numpy()
